+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Upcoming
+
+### Added
+
+- Added PostCSS and Tailwind
+
+### Changed
+
+- Converted CSS plugin to PostCSS in `rollup.config.mjs`
+
 ## 1.1.0 - 2024-08-31
 
 ### Added