+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Upcoming
 
+## 1.1.1 - 2024-10-9
+
 ### Added
 
 - Added PostCSS (`postcss.config.js`)